Bug reported by Alexandre Fruchaud
Imagine you have three nodes: 101, 102, 103.
101 becomes leader.
101 replicates log to 102, 103.
Following a technical issue, 103 is isolated from the network.
103 failure detector notices that 101 is unreachable and 103 becomes a candidate. 103's term is increased to 2.
103 can't be elected leader (it only gets it's own vote) and becomes stuck.
103 election's timeouts. 103's term is increased to 3.
103 comes back into the network.
101 is broadcasted new data, which it appends to it's log.
102 receives a log request from 101, and accepts it.
103 receives a log request from 101, but rejects it since the term of 103 > the term of 101 (which is 1).
103 election's timeouts. 103's term is increased to 4.
103 sends a vote request to 101 and 102.
101 and 102 both deny the request since 101 and 102 log's length > 103 log's length.
103 election's timeouts. 103's term is increased to 5.

In this case, I don't think it's possible for 103 to recover from the network isolation.

Just for you know I've been studying in the meanwhile the TLA+ proof of Raft
(https://github.com/ongardie/raft.tla)
In their spec, any node n1 that receives a request from another node n2 with a greater term than
them makes them update their term such as n1.term = n2.term and resetting n2 role as follower (with
no leader). An election would happen, and nobody would vote for n1 (in our case 103) because n1 log
is outdated, so it would be a guarantee that 101 or 102 is elected leader.