Revert "Easier-to-read formulation of the CommitLogEntries function"
This reverts commit 21801be478550325625a0d12cb96d94e72fa7504.

Jorn Janneck pointed out that my modified version of CommitLogEntries is
not safe, because it allows a message to be committed if it was
replicated to a quorum of nodes in a previous epoch. The Raft paper
explains in Section 5.4.2 and Figure 8 why it's not safe to commit a
message in these circumstances (it's possible for the message at that
log index to change if leadership changes).

Reverting to my previous pseudocode, which I think accurately reflected
the Raft TLA+ spec.